Move `ArticleEmbedder` to its own file
