+++ src/poprox_recommender/embedders/__init__.py
@@ -1,0 +1,3 @@
+from poprox_recommender.embedders.article import ArticleEmbedder
+
+__all__ = ["ArticleEmbedder"]
